requirements.txt
flask
google-cloud-recommender
google-cloud-asset
google-cloud-policy-troubleshooter